gremtech/requirements.txt
apturl===0.5.2ubuntu4
Brlapi==0.6.1
chardet==2.2.1
checkbox-ng==0.3
checkbox-support==0.2
command-not-found==0.3
defer==1.0.6
Django==1.9
django-debug-toolbar==1.4
django-ipware==1.1.5
django-neue-transmeta==0.8.0
feedparser==5.1.3
friends==0.1
httplib2==0.8
language-selector==0.1
louis==2.5.3
lxml==3.3.3
Mako==0.9.1
MarkupSafe==0.18
oauthlib==0.6.1
onboard==1.0.1
oneconf==0.3.7
pep8==1.7.0
Pillow==3.1.1
piston-mini-client==0.7.5
plainbox==0.5.3
pycrypto==2.6.1
pycurl==7.19.3
pygobject==3.12.0
pyparsing==2.0.1
python-apt===0.9.3.5ubuntu2
python-debian===0.1.21-nmu2ubuntu2
pytz==2015.7
pyxdg==0.25
PyYAML==3.11
requests==2.2.1
six==1.5.2
software-center-aptd-plugins==0.0.0
solid-i18n==1.2.0
sqlparse==0.1.18
ubuntu-drivers-common==0.0.0
ufw===0.34-rc-0ubuntu2
unattended-upgrades==0.1
unity-scope-audacious==0.1
unity-scope-calculator==0.1
unity-scope-chromiumbookmarks==0.1
unity-scope-clementine==0.1
unity-scope-colourlovers==0.1
unity-scope-devhelp==0.1
unity-scope-firefoxbookmarks==0.1
unity-scope-gdrive==0.7
unity-scope-gmusicbrowser==0.1
unity-scope-gourmet==0.1
unity-scope-guayadeque==0.1
unity-scope-manpages==0.1
unity-scope-musique==0.1
unity-scope-openclipart==0.1
unity-scope-texdoc==0.1
unity-scope-tomboy==0.1
unity-scope-virtualbox==0.1
unity-scope-yelp==0.1
unity-scope-zotero==0.1
urllib3==1.7.1
usb-creator==0.2.23
xdiagnose===3.6.3build2
xkit==0.0.0
